requirements.txt
aiohttp==3.6.2
numpy==1.18.2
discord.py==1.3.4
discord==1.0.1
Pillow==7.1.1
python-dotenv==0.13.0
PyYAML==5.3.1
praw==7.1.0
